pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "transaction-analysis"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv==1.0.1",
    "click==8.1.7",
]

[project.scripts]
ta = "transaction_analysis.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
